backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
joblib
scikit-learn
pydantic
python-dotenv
